chore(cmd/grpc): rename pushShutdown onShutdown and move clientConn to main
